+++ d60.Cirqus.MongoDb/Snapshotting/NewSnapshottingAggregateRootRepositoryDecorator.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using d60.Cirqus.Aggregates;
 using d60.Cirqus.Events;
@@ -69,7 +68,7 @@ namespace d60.Cirqus.MongoDb.Snapshotting
             var aggregateRootInstance = _aggregateRootRepository.Get<TAggregateRoot>(aggregateRootId, unitOfWork, maxGlobalSequenceNumber, createIfNotExists);
             var checkedOutSequenceNumber = new AggregateRootInfo(aggregateRootInstance).SequenceNumber;
 
-            SaveSnapshot(aggregateRootId, aggregateRootInstance, checkedOutSequenceNumber, snapshotAttribute, maxGlobalSequenceNumber);
+            SaveSnapshot(aggregateRootId, aggregateRootInstance, checkedOutSequenceNumber, snapshotAttribute, maxGlobalSequenceNumber, false);
 
             OnCommitted(aggregateRootId, unitOfWork, aggregateRootInstance, checkedOutSequenceNumber, snapshotAttribute);
 
@@ -96,26 +95,27 @@ namespace d60.Cirqus.MongoDb.Snapshotting
         {
             unitOfWork.Committed += eventBatch =>
             {
-                SaveSnapshot(aggregateRootId, aggregateRootInstance, checkedOutSequenceNumber, snapshotAttribute, eventBatch.Max(e => e.GetGlobalSequenceNumber()));
+                SaveSnapshot(aggregateRootId, aggregateRootInstance, checkedOutSequenceNumber, snapshotAttribute, eventBatch.Max(e => e.GetGlobalSequenceNumber()), true);
             };
         }
 
-        void SaveSnapshot(string aggregateRootId, AggregateRoot aggregateRootInstance, long checkedOutSequenceNumber, EnableSnapshotsAttribute snapshotAttribute, long validFromGlobalSequenceNumber)
+        void SaveSnapshot(string aggregateRootId, AggregateRoot aggregateRootInstance, long checkedOutSequenceNumber, EnableSnapshotsAttribute snapshotAttribute, long validFromGlobalSequenceNumber, bool instanceIsBasedOnSnapshot)
         {
             var info = new AggregateRootInfo(aggregateRootInstance);
             var currentSequenceNumber = info.SequenceNumber;
-            if (currentSequenceNumber == checkedOutSequenceNumber) return;
+            if (instanceIsBasedOnSnapshot && currentSequenceNumber == checkedOutSequenceNumber) return;
 
             var serializedInstance = _sturdylizer.SerializeObject(info.Instance);
 
-            _snapshots.Save(new NewSnapshot
+            _snapshots.Insert(new NewSnapshot
             {
                 Id = string.Format("{0}/{1}", aggregateRootId, currentSequenceNumber),
                 AggregateRootId = aggregateRootId,
                 Data = serializedInstance,
                 ValidFromGlobalSequenceNumber = validFromGlobalSequenceNumber,
                 Version = snapshotAttribute.Version
-            });
+            },
+            WriteConcern.Unacknowledged);
         }
 
         AggregateRoot PrepareFromSnapshot(NewSnapshot matchingSnapshot, long maxGlobalSequenceNumber, IUnitOfWork unitOfWork)